Renamed WidgetRegistry
